cheat to support expression clauses at any position.
it's inefficient, but it works for now.

a bit farther out, the question is: do we want to actually "understand"
the queries (dependencies, clauses that can be applied) or do we want to
continue doing this on the current level? (e.g. we could also sort the
clauses such that the expression clauses occur right after all vars in
it are bound.)
not sure, understanding the clauses would making optimizing them
easier/possible, but might also make functions that don't care about
that more complex. mhh...