support for querying multiple dbs.
this allows pushing the dbs in the environment and not having to pass
the dbs around. though maybe we may want to put the dbs in a separate
map and pass that around.
mhh, maybe i'll do that. (having the dbs in the env makes debugging it
kind of annoying.)