compare the suspicious slow query.
surprisingly it's not the clause ordering that causes the slowdown but
because we put the result in a set only after the query.

(that the different is so significant (10x speed up without putting the
result in a set) might be due to the small dataset. well, i kind of
expect that.)