better support for recursive rules.
phew. they key part is remembering which paths rules have already been
evaluated. everything else was already kind-of qsqr. i'm not sure
whether this is already it, but for now it's pretty good.

if you're curious the explanation of qsqr in "foundations of databases" [1]
was the most helpful, but this paper [2] also helped a bit.

i'm not sure whether a struggle this long is usual or just me being
unfamilar with reading papers, taking not enough time for reading &
thinking or whether in the end trying to write about it was best.

anyway, it's nice to be at this point now.

onwards! :)

[1]: http://webdam.inria.fr/Alice/pdfs/all.pdf
[2]: https://domino.mpi-inf.mpg.de/intranet/ag5/ag5publ.nsf/0/5132110d0a14b0fbc12578f000418a98/$FILE/d2r2-ruleml.pdf