wrap the rule state in an atom.
this prevents us from evaluating recursive rules with cycles in
different "branches" more than once. in some cases it's about twice as
fast as not passing it.
we probably could have passed the state back with the result, but i
think using an atom is both simpler and easier in this case. (is this
"idiomatic"? are there drawbacks to this?)